+++ packages/collector/src/immediate.js
@@ -6,8 +6,8 @@ const { util: coreUtil } = require('@instana/core');
 // modifying the source code. See
 // https://www.instana.com/docs/ecosystem/node-js/installation#installation-without-modifying-the-source-code
 
-let excludedFromInstrumentation = coreUtil.excludedFromInstrumentation && coreUtil.excludedFromInstrumentation();
+let isExcludedFromInstrumentation = coreUtil.excludedFromInstrumentation && coreUtil.excludedFromInstrumentation();
 
-if (!excludedFromInstrumentation) {
+if (!isExcludedFromInstrumentation) {
   require('./index')();
 }